app/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
b2sdk==1.23.0
bidict==0.22.1
binaryornot==0.4.4
boto3==1.28.36
botocore==1.31.36
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.3.0
cssbeautifier==1.14.9
Django==4.2.4
django-backblaze-b2==4.0.2
django-browser-reload==1.11.0
django-ckeditor==6.7.0
django-htmx==1.16.0
django-js-asset==2.1.0
django-quill-editor==0.1.40
django-storages==1.13.2
django-tailwind==3.6.0
djlint==1.32.1
EditorConfig==0.12.3
Faker==19.3.0
fastapi==0.101.1
fastapi-socketio==0.0.10
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
htmx==0.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
logfury==1.0.1
loguru==0.7.0
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
nicegui==1.3.9
orjson==3.9.4
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
playwright==1.37.0
pluggy==1.2.0
pscript==0.7.7
psycopg2-binary==2.9.9
pydantic==1.10.12
pyee==9.0.4
Pygments==2.16.1
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.8.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
